requirements.txt
aiofiles==24.1.0
aiogram==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
black==25.1.0
certifi==2025.8.3
click==8.2.1
distro==1.9.0
docstring-to-markdown==0.17
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
jedi==0.19.2
jiter==0.10.0
magic-filter==1.0.12
multidict==6.6.4
mypy_extensions==1.1.0
ngrok==1.5.1
nodeenv==1.9.1
openai==1.100.1
packaging==25.0
parso==0.8.4
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
propcache==0.3.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyngrok==7.3.0
pyright==1.1.403
python-dotenv==1.1.1
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.13.0
python-telegram-bot==22.3
PyYAML==6.0.2
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
ujson==5.10.0
yarl==1.20.1
zipp==3.23.0
